Put transformTile, transformPoint in separate file
